Extracts plus tests from the eval tests
