+++ partiql-eval/src/test/kotlin/org/partiql/eval/PTestCase.kt
@@ -1,0 +1,8 @@
+package org.partiql.eval
+
+interface PTestCase : Runnable {
+    /**
+     * Executes the test case
+     */
+    override fun run()
+}
